Switched to STL vectors.
No point reimplementing resizing logic everywhere, vectors are not too
painful over object pointers as it defeats a lot of the semantic pain.

Problem with the type hierarchy. It seems logical to define a base class
for a "generic" type and then derive instances from it for the base
classes in the IR type-system. This cannot work as inside the Value or
Instruction class we need the least general union (ie it has been upcasted
to a Type), so then we do not have the more specific information about
what type it is. Overloading equality methods won't work as the type
info is static rather than dynamic. We could use RTTI but it would get
messy as we need to despatch from the LGU to the most specific derived
class in each case...

Trying to encode this inside the C++ type system as a class hierarchy will
create a huge ball of crap, but the other way is to encode it as data inside
the Type class which implies fixing the (target language) type system...