Added Values
Types are now constructed from a string and a size, the size is then accessible
as a read-only member. The Value object is now exposed to Python and the method
Block.input() is correctly linked to use these objects for input and output.

The interface for building a type is directly in the pycodestew module rather
than inside SimpleMachine. This may make sense for the C++ code as well.