Reorganising into separate Machines.
Introduced two new classes to represent code for the x86 and arm machines.
It has seemed debatable whether these should be classes or just functions
that convert a Block of SimpleMachine operations into a string for output.

The decision is because there will be non-trivial graph transformations
during the conversion and so we will need to be able to output the result
for comparison. The x86 machine will need to coalesce values for 2-op
instructions, inserting MOV nodes as appropriate. Both machines will need
to handle conversion of UBITS of non-register width; i.e splitting and
synthesizing instructions as appropriate. This moves a non-trivial chunk
of the CAO code transformation from the middle into this back-end and
allows code in a SimpleMachine to be retargetted much more effectively at
different output machines.

Reorganised the Makefile to store/make objects for each module.