Added Python interface for Instructions and topSort().
This highlights an interesting problem, reference equality in the C++
code will not match reference equality at the Python level. We
definitely want equality over references at the Python level to correctly
match uniqueness of objects. This problem arises because the complex
structure that we are sharing between the language is exported in a
fine-grained manner to give the Python code access to edges and vertices
rather than just a graph black-box.

The dense ref space within each block suggests a partial solution, but
if we try to implement sharing (via PyObject caching) at this level
then we will need to hold refs on objects, and this may bugger up GC
as we can't know when things die on the outside. Have to check out the
dealloc interface...