Finished off add128 translation.
Missing value was the middle word of the output, needed to check the ubits
sizes and handle the difference between carry spilling and discarding. The
carry flags are represented explicitly in the graph translation as one-bit
values - these encode the control dependencies between the addc instructions.
The final spill is represented as a signext instruction from ubits<1> to ubits<64>.

The trivial allocator should still handle this block after explicit support
for "car" as a register is added.