Basic proof of concept for x86 register allocation.
This implements the trivial allocation for x86 accounting for some of
the issues in the 2-op instruction set. Values that are defined by
instructions that coalesce their output are set to implied, a second
pass will copy the input register onto the target. The other issue to
account for is that inputs to coalescing operations must have copy
instructions inserted to clone the value if it has more than one use.