Real code generation from stored allocation.
The register allocation is now represented by a class that extends the
basic block. The clone() method has been tweaked a bit so that initialising
an Allocation from a Block via cloning is possible.

The Projection now sets the inputs on the target block to reflect the set
of words that originated from the inputs on the source block. This allows
valid topSorts to exist so that we can process the block.

The first two addition instructions are generated correctly. Carry spills
and the sign extension will be a little more tricky.