Carry spills through signext instruction.
This was easier than anticipated. We could probably formalise what
makes a super-instruction in terms of graph vertices that are
instantiated as sequences of real instructions. If no scratch
registers are needed and the only register effects are on the
declared outputs of the vertex then it seems quite easy.