First step in getting the ARM backend to compile add128.
This is quite messy as the code to handle expansion of add() instructions
is cut and paste into the arm backend. Seems reasonably generic so should
probably be shared somewhere between the two.

No register allocator is plugged in yet, but there should be an analogue
of the trivial allocator for the ARM register set.