First stab at a modulo scheduler in the ARM backend.
Modulo is the wrong terminology; this isn't allocating registers
according to a fixed period across loop iterations. This is executing
a simulation of the graph and using the live/dead state to allocate
the registers.

The debug looks like it works, but the graphviz output has no strings
in the registers so something goes wonky...