Valid ARM and x86 code for the add128 cogen.
Setting the carry register in a pre-phase causes a problem that we can't
see yet. The graph traversal cannot cope with pre-allocated registers.
This doesn't show up yet as we do not attempt to hold the carry register
in the live/free sets, although when the MUL instruction is added for
the x86 target it will become a problem.

The other issue is that the carry chains cannot be split, this doesn't
show up in this simple example but if we have two independent chains
then we cannot multiplex the carry flag between them. This will need
some other form of dependencies in the instruction scheduling that are
stronger than just a valid topological sort.