Reorganisation.
The old add128 cogen is now replaced by the commonMain / add128 combo.
The project now uses a build/ and a src/ directory, cleaning out the
junk in the root. Currently the cogens are built in the cogen directory,
but this is a bit of a pain as the source modules live there and the
mac version creates a .dSYM/ folder which needs to be cleaned out.

Currently the ARM test cases go through (although the asm result ends
up on the console) and the x86 cases still need to be tweaked.