Working results for two testcases.
Using stems in the Makefile is much simpler than rebuilding target
names from subst functions. Added error recovery in the following
sense: default "error" outputs are created before the cogen
execution, and the error code is ignored by make so that each of the
testcases now forms an independent experiment and continues when
other testcases fail.

The add64/x86 testcase currently fails with an error.