Makefile driven by testcases
Flipped the makefile around, now there is a list of test-cases that is used
to generate the lists of dot-files and pdfs for output. Fixed the dot-files
being removed as intemediate results by make.

Expanded the set of testcases to 3x for add and xor on both architectures.
Changed the failure handling so that errors are pushed out to the commonMain
and every failure path terminates the log with "Failure" while every
successful path terminates the log with "Success".

These changes allow the execution of all of the experiment testcases generating
a simple progress chart for whether or not the cogen successfully produced
output. Currently the xor cases all fail as the projection is not implemented,
and the 64-bit x86 case breaks for addition as there is a strange glitch in
the projected intermediate code.