Final changes for machine reference.
Every block in the system should now be associated with a machine. The cloning
interface is a bit weird (implemented as copyInto rather than a copy
constructor). This function now destroys any pervious data in the block, and
the Allocation constructor needs to call its super to setup the block ref.