Reworked the add projection.
Now that the input is firmed up it is quite simple to convert the addition
projection to something that works in the single word case. After the changes
are made it is more obvious that this piece of code is a simple interpreter
for addition of carry-chains, recording the stateful effect of the spilled
carry into the resulting graph. The more specific instruction set than the
concrete machine records the necessary graph structure, instead of the wider
assumption that carries are always propagated. This can be seen as a unique
formulation between the input machine (addition over pure bit-strings) and
the output machine (just uses add/adc) which destroys this information. All
of the addition cases now succeed.