Refactored the add projection.
Building up instructions by individual calls to add values leads to
messy and hard to follow code. Added helper functions that build
specific types of instructions in specific machines so that the
value input/output patterns can be expressed in the helper function
arguments.

It is not possible to hold these helper functions in the Block class
as it is generic, and introducing a hierarchy of X86Block, ArmBlock
etc seems clumsy. Instead the helper functions have been introduced
into the Machine class hierarchy, in the long-term the similiarities
between the X86Machine and ArmMachine can merged into a shared
super-class.

For now this has pulled the translation function into copies in both
of the Machine classes, although this duplication is messy it allows
a much simpler formulation of the translation using the helper functions.