Two new testcases.
Bigger inputs sizes to trigger failure in the register allocator. The
failure cases are not quite as desired - the input set is to large to
fit in registers making an ABI impossible. Next thing to try is a testcase
with a larger working-set : input-set ratio. This will be a multiplication
instruction.