New testcase.
MD5 is interesting, although it is very broken and so not useful in the
real world it is an example of a program just outside the set of currently
expressible progrms. To make this compile will require several new additions,
several new instructions in the SimpleMachine, followed by some new
translations to convert this into the concrete machines.

This is not the next step, next will be some simpler multiplications and
then MD5 as a precursor to the gamit of hash functions, block ciphers and
stream ciphers.