Graphviz breaks down on the md5 src program.
We can't really rely on Graphviz. It should be able to handle these graphs
as they only contain a couple of thousand vertices but it triggers some
kind of failure case.

As the class of graphs that need to be visualised is very restricted in
comparison to the Graphviz input I've started writing a domain specific
replacement. The first thing to do is to simplify the layout problem
using structural similiarity. A basic principle in graph layout is that
similarity in the graph should be reflected by symmetry in the result.

To manipulate induced subgraphs I've started working on a Region class that
indicates a marked subset of vertices (and any induced edges). This is
used to partition the graph in a bottom-up process that splits the
graph into partitions according to an equality relation that considers
local graph structure. It is computing an indirection representation of
isomorphisms amongst the sub-graphs (the indirect part is a numbering of
the vertices within the subgroup that preserves the relation).