Implemented the executability relation on the RegionX.
After implementing this graph search algorithm for about the billionth
(and hopefully the last time) it occurs to me that it has never been
published. Really should write a tech report about how this generalises
the dominance relation over bi-graphs and its connection to reachability
and valid simulations of a dataflow program.