Initial views of MD5
Fixed a couple of bugs: the horrific mess in RegionX::repr() was overflowing
and corrupting memory. Replaced it with something more stable. Typo in the
opcode enum meant that the extract and lor ops had the same value. Added some
more code for outputting the Regions / Blocks.

Now that GraphViz can render a partial view of MD5 the small size of the
dominated regions is obvious: there is a huge degree of diffusion in the
hash function which shows up as a single dependency chain with frequent
inputs from earlier values in the chain and the very wide extraction of
words from the input chunk.

The extraction process is essentially parallel so it produces a very wide
graph (everything has the same rank). The computation process is very
sequential (single chain). The edges from the wide single rank to almost
every rank in the chain are as randomly distributed as possible. Explains
why GraphViz chokes on the input...

Next step is returning to computing isomorphisms between subgraphs. Although
subgraph isomorphism is hard in general this graph has highly distinct local
structure so it is far from the worst case. Inside the chain the nodes with
high fan-out are easy to spot and should be isomorphic to one another.