Splitting collections into partition based on local equality.
New module contains the bin-building code generalised over a template.
The first use is splitting the Values in the Block into bins based on
an equality relation that takes into account the number of uses (out-degree)
and the type. The next step is looking at the uses as a vector of
Instructions and re-partitioning based on the local equivalency relation.
This currently checks that the opcodes match and each input/output operand
has the same type.

If the shape of the use-partition matches then the partition of the data-values
stands, otherwise it can be split into smaller bins where the using instructions
differ. The entry point in the isomorphism module should throw away
"uninteresting" bins (judged by size of graph regions in each element and the
number of elements in the bin) and scan each remaining bin to a higher
depth.

At some point the bins within each bin need to be resolved, which is very
expensive. Either all permutations need to be checked or some kind of
backtracking implemented. The distribution of graph regions to bins is
essentially an encoding of the constraints entailed by the regions matching
other elements in the same bin.