Rough sketch of the somethings...
Each of the bins is now stored in a container called a CanonSet,
within which each Canon contains separated sets of graph vertices.
The Values referenced are assumed to be in a definite order. The
rationale for this is that we will hang each isomorphic region
from a root Value and the process of binding them together will
establish an (arbitrary but consistent) ordering for them. The
Instructions are split into definites and possibles. The definites
have unique enough traces from the root that we can distinguish
between them, and thus we know the ordering is preserved across
every Canon in a CanonSet. The possibles are a partition of the
remaining Instructions according to a local equivalence relation.
Equality is preserved over blocks of these partitions between
Canons in a CanonSet, but the exact ordering is unknown. If two
Canons refer to isomorphic regions then some unknown permutation
of these sets will produce the isomorphic mapping between them.

This generates the two-level binning as before, then splits the
bins according to the partition of the possibles partition. Bins
2 and 3 are currently split incorrectly when they should remain
merged. The reason is that the when we compare the partitions we
treat them as distributions of graph sructure and compare them
by shape - but the blocks within the partition have not been
ordered so the comparison is invalid. Any consistent ordering will
do for the blocks - opcode ordering for example.