Colourised regions across Isomorphisms.
We now generate a Region for each of the matching vertices across
an Isomorphism. So one Region contains the same vertex from each
Canon in the Isomorphism. These regions are coloured separately
in the dot output to show how the structure maps across.

Next up, the possibles partition needs to be replicated for each
value in the Canon.