Back to a working state.
There are differences in the equality relations that change how the
structure unfolds on the graph. Currently the code renders a hardcoded
isomorphism and it was looking at the wrong one. This is now hardcoded
for the right one. This will change again next some extra thresholding
and splitting steps are required.