Splitting the add instructions.
The add vertices below the heads of the mains sections have the easiest
structure to separate. One of them leads into the high-degree head of
the next section and one leads into a low-degree tree-like region.

Extracting the blocks of add instructions from the possibles partitions
is essentially a slicing operation over a 3d array. As long as the
Isomorphism is in a legal state the sub-structure is entirely regular so
this could be refactored into an explicit slicing operation.

The blocks of possible instruction are projects onto blocks of produced
values. These blocks are then compared for differences. The result is a
partition of the blocks that may remove some of the IsoRegions from the
Isomorphism (in this case there are structural differences half-way
through the outermost MD5 loop and in the final iteration). Within the
largest matching block each of the Values has been ordered according to
locally visible differences.

There needs to be a check that the order is strictly increasing/decreasing
as adjacent neighbours cannot be distinguished and must remain in the
possibles vector after processing. For distinguished values we can project
the results back onto the vector of instructions to produce an order that
is consistent across the Isomorphism. This will allow those vertices to be
confirmed (pushed into definites and produced values make new section heads).