Adds now split correctly and confirmed.
After the partition of the defined values the split is mapped back onto
the instructions that produced them within the blocks of the possibles
partition inside each section. There are several locate() methods for
finding instruction indices that could be tidied up later.

The defined values are not propagated as boundary conditions needs to be
handled before we push any further. Most of the main region of the hash
function is now split into IsoRegions. Further propagation will require
classification of vertices to prevent overlap.