Started refactoring the confirm code.
The confirm function in the Section has been split into smaller
pieces so that it can be called from incProductions. The confirm
function in the IsoRegion has been updated.

There are currently several pieces of code that work on undocumented
assumptions. I've started to document the assumptions for the confirm
functions. This is a new language construct, pieces of code that
operate on independent pieces of data (similar to vectorisation),
but under the assumption that the data is isomorphic across some set
the loop iterates over. These pieces of code are isomorphism
preserving across that set.