First full attempt at simplifying incProductions.
The function is still about the same length but now it is mostly comments
and much less code. Started to fold out the common operations over the
structures like slicing and mapping. Still needs some more simplification
before we can loop over each possibles block and drive the process based
on the classification of the new Section heads.

Added a Partition class. The partition function will be deprecated and
calls rewritten to use the class.