Buggy version
Found a way to clean up the incProductions drastically. The old approach
was hiding a mismatch between assumptions on the structure of the
possibles partition and what was being calculated. Documenting the
assumptions and preconditions of the calls has revealed this. Essentially
the Isomorphism/IsoRegion/Section/Possibles structure did not have to
be as regular as it was as it can diverge and then be split later through
a distinguish() call. This was masked by the horrible four-level access.

New version does not work yet as there is an old bug causing problems.
I had though the blank lines in dump() were just a glitch in displaying
the data but they are actually empty blocks in the possibles partition
being created earlier on. Need to track this down.