Fixed the empty block bug.
When the instruction was erased from the block in Section::confirm
no check was made to see if it was the last instruction in the
block.

The output in Graphviz looks superficially like the result of the
old incProductions() call, but the dump of the regions shows that
more merging has occured and the regions now overlap. This was
going to be the next stage after the old code.

Vertices need to be classified into { ThisRegion | OtherRegion | Outside }
and only pushed into section heads when overlap will not occur.