Folding code handles inputs.
The regions are now correctly identified and fold instructions are inserted
into the graph. Edges that cross into the region are moved so that they
target the new instruction (in the correct operand slot) instead of the
old vertex.

This makes areas of the graph unreachable so the old visualisation code has
been disabled and the entire graph is now output, manually cropped in the
dot file to remove structure and then the result is cropped using gs to
make it visible under linux.