Outputs for fold regions as well.
It is now obvious that there are four edge crossing cases to consider rather
than two. When a Value inside the region has an edge connected to an instruction
outside the region it breaks the bi-partite property of the graph.

To fix this the I/O values of the region need to be cloned "shadowed" by Values
in the outer graph. As these Values will have edges to/from the input/output operands
of the fold instruction this will preserve the property in the graph.